da Float a Double

+++ src/main/java/it/unipi/dii/ingin/lsmsd/fantamanager/page_controllers/SeeCardController.java
@@ -154,9 +154,9 @@ public class SeeCardController implements Initializable{
 
     public void buy_card(MouseEvent mouseEvent) throws NoSuchAlgorithmException {
 
-		double credits=card_doc.getDouble("credits");
+		int credits=card_doc.getInteger("credits");
 		if(check_credits((int)credits)) {
-			OptionsMongoDriver.update_user_credits(false, global.user.username, (int) credits);
+			OptionsMongoDriver.update_user_credits(false, global.user.username, credits);
 			card_collection bought_card = new card_collection(card_doc.getInteger("player_id"), card_doc.getString("fullname"), 1, card_doc.getString("team"), card_doc.getString("position"));
 			collection.add_card_to_collection(bought_card, global.id_user);
 			System.out.println("Card bought");
